pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "moltbook-toolkit"
version = "1.0.0"
description = "Security-first toolkit for building AI agents on Moltbook"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Nir Diamant", email = "diamantnir@gmail.com"}
]
maintainers = [
    {name = "Nir Diamant", email = "diamantnir@gmail.com"}
]
keywords = [
    "moltbook",
    "ai-agents",
    "llm",
    "security",
    "prompt-injection",
    "openclaw",
    "social-network",
    "ai-safety"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "requests>=2.28.0",
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]
all = [
    "moltbook-toolkit[dev,docs]",
]

[project.urls]
Homepage = "https://github.com/NirDiamant/Building_Moltbook_Agents"
Documentation = "https://github.com/NirDiamant/Building_Moltbook_Agents#readme"
Repository = "https://github.com/NirDiamant/Building_Moltbook_Agents.git"
Issues = "https://github.com/NirDiamant/Building_Moltbook_Agents/issues"
Changelog = "https://github.com/NirDiamant/Building_Moltbook_Agents/releases"

[project.scripts]
moltbook = "tools.moltbook_cli.cli:main"

[tool.setuptools]
packages = ["tools", "tools.moltbook_cli", "tools.moltbook_cli.commands"]
package-dir = {"" = "."}

[tool.setuptools.package-data]
"*" = ["*.md", "*.json", "*.yml", "*.yaml"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.venv
    | build
    | dist
)/
'''

[tool.ruff]
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=tools --cov-report=term-missing"
